added entity control logic

+++ src/main/java/com/example/kadastr/service/impl/CommentServiceImpl.java
@@ -3,9 +3,12 @@ package com.example.kadastr.service.impl;
 import com.example.kadastr.dao.CommentDAO;
 import com.example.kadastr.dao.NewsDAO;
 import com.example.kadastr.dto.CommentDto;
+import com.example.kadastr.exception.AuthException;
+import com.example.kadastr.exception.IllegalControlException;
 import com.example.kadastr.exception.NoSuchIdException;
 import com.example.kadastr.model.Comment;
 import com.example.kadastr.model.News;
+import com.example.kadastr.security.util.AuthHelper;
 import com.example.kadastr.service.CommentService;
 import com.example.kadastr.util.Mapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,12 +25,14 @@ public class CommentServiceImpl implements CommentService {
     private final CommentDAO commentDAO;
     private final NewsDAO newsDAO;
     private final Mapper<Comment, CommentDto> commentMapper;
+    private final AuthHelper authHelper;
 
     @Autowired
-    public CommentServiceImpl(CommentDAO commentDAO, NewsDAO newsDAO, Mapper<Comment, CommentDto> commentMapper) {
+    public CommentServiceImpl(CommentDAO commentDAO, NewsDAO newsDAO, Mapper<Comment, CommentDto> commentMapper, AuthHelper authHelper) {
         this.commentDAO = commentDAO;
         this.newsDAO = newsDAO;
         this.commentMapper = commentMapper;
+        this.authHelper = authHelper;
     }
 
     @Transactional(rollbackFor = Exception.class)
@@ -44,12 +49,13 @@ public class CommentServiceImpl implements CommentService {
 
     @Transactional(rollbackFor = Exception.class)
     @Override
-    public void createComment(UUID newsId, CommentDto commentDto) throws NoSuchIdException {
+    public void createComment(UUID newsId, CommentDto commentDto) throws NoSuchIdException, AuthException {
         News news = newsDAO.findById(newsId)
                 .orElseThrow(
                         () -> new NoSuchIdException("News with uuid = " + newsId + " doesn't exist")
                 );
         Comment comment = commentMapper.mapToModel(commentDto);
+        comment.setInsertedById(authHelper.getCurrentUserUUID());
         comment.setNews(news);
         comment.setIdNews(news.getUuid());
         news.getComments().add(comment);
@@ -58,17 +64,21 @@ public class CommentServiceImpl implements CommentService {
 
     @Transactional(rollbackFor = Exception.class)
     @Override
-    public void updateComment(UUID uuid, CommentDto commentDto) throws NoSuchIdException {
+    public void updateComment(UUID uuid, CommentDto commentDto) throws NoSuchIdException, IllegalControlException {
         Comment comment = getCommentModelById(uuid);
-        comment.setText(commentDto.getText());
-        commentDAO.save(comment);
+        if (authHelper.checkUserControlOnEntity(comment)) {
+            comment.setText(commentDto.getText());
+            commentDAO.save(comment);
+        }
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
-    public void deleteComment(UUID uuid) throws NoSuchIdException {
-        getCommentModelById(uuid);
-        commentDAO.deleteById(uuid);
+    public void deleteComment(UUID uuid) throws NoSuchIdException, IllegalControlException {
+        Comment comment = getCommentModelById(uuid);
+        if (authHelper.checkUserControlOnEntity(comment)) {
+            commentDAO.deleteById(uuid);
+        }
     }
 
     private Comment getCommentModelById(UUID uuid) throws NoSuchIdException {
